requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.7.0
email_validator==2.2.0
Flask==2.2.5
flask-cors==5.0.1
flask-mongoengine==1.0.0
Flask-WTF==1.2.2
google_recaptcha_flask==0.1.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
mongoengine==0.29.1
more-itertools==10.6.0
packaging==24.2
premailer==3.10.0
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.2.1
yagmail==0.15.293
